+++ src/app/calculator/furnaces/furnaces.component.html
@@ -52,7 +52,7 @@
             </p>
           </div>
         </div>
-        <div class="row calc-item-row">
+        <!-- <div class="row calc-item-row">
           <div class="col-1">
             <img src="assets/images/assessment.png" class="ico-size2">
           </div>
@@ -65,86 +65,26 @@
               improvements.
             </p>
           </div>
-        </div>
+        </div> -->
       </div>
     </div>
   </div>
 </div>
 
 <div class="stand-alone-calc tab-content" *ngIf="getSelectedTool() == 'o2-enrichment'">
-  <!-- <div class="row">
-    <div class="col-lg-11 mx-auto">
-      <h2>Fuel Consumption</h2>
-      <p>Calculation of available heat (an indication of thermal efficiency) for fuel fired furnaces and expected energy savings
-        when oxygen in combustion air, is changed from standard (21%) to a higher value. Can add multiple lines to the graph
-        by changing values under "Combustion with Oxygen enriched Air" and clicking "Plot".
-        Changing values for "Combustion with Air" will adjust all the lines on the graph.</p>
-    </div>
-  </div> -->
-
   <app-o2-enrichment></app-o2-enrichment>
 </div>
 
 
 <div class="stand-alone-calc tab-content" *ngIf="getSelectedTool() == 'efficiency-improvement'">
-  <!-- <div class="row">
-    <div class="col-lg-11 mx-auto">
-      <h2>Efficiency Improvement Calculator</h2>
-      <p> Provides calculation of available heat (the heat that remains within the heating system as a fraction of the heat input,
-        which is an indication of thermal efficiency) and fuel savings for fuel fired furnaces and expected energy savings
-        when the burner operating conditions are changed. Achieving a decrease in flue gas oxygen implies a reduction in
-        excess air, which increases the available heat. Enter data for “current” and “new” configurations.
-      </p>
-    </div>
-  </div>
-  <div class="calculator-container"> -->
   <app-efficiency-improvement></app-efficiency-improvement>
-  <!-- </div> -->
 </div>
 
 
 <div class="stand-alone-calc tab-content" *ngIf="getSelectedTool() == 'energy-equivalency'">
-  <!-- <div class="row">
-    <div class="col-lg-11 mx-auto">
-      <h2>Energy Equivalency Calculator</h2>
-      <p> Provides calculations for converting values of energy input requirements when the heat source is changed from fuel-firing
-        (Btu/hr or kJ/hr) to electricity (kW), or vice versa.</p>
-    </div>
-  </div>
-  <div class="calculator-container"> -->
   <app-energy-equivalency></app-energy-equivalency>
-  <!-- </div> -->
 </div>
 
 <div class="stand-alone-calc tab-content" *ngIf="getSelectedTool() == 'energy-use'">
-  <!-- <div class="row">
-    <div class="col-lg-11 mx-auto">
-      <h2>Flow Calculations Energy Use</h2>
-      <p> Calculation of flow rates for commonly used gases using data for an orifice type flow meter and heat input calculations
-        when the gas is a fuel (i.e. natural gas, propane).</p>
-    </div>
-  </div> -->
-
   <app-energy-use></app-energy-use>
-
-</div>
-
-
-<div class="stand-alone-calc tab-content" *ngIf="getSelectedTool() == 'pre-assessment'">
-  <!-- <div class="row">
-    <div class="col-lg-11 mx-auto">
-      <h2>Pre-assessment / Screening</h2>
-      <p>This calculator allows the user to enter either energy use information gathered from meters or design specifications
-        (for fuel, electrotechnology or steam furnaces) to calculate energy use for multiple furnaces. This allows users to
-        compare the relative energy use of furnaces within a facility to prioritize energy evaluations and efficiency improvements.
-      </p>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-lg-11 mx-auto">
-      <div class="calculator-container"> -->
-  <app-pre-assessment></app-pre-assessment>
-  <!-- </div>
-    </div>
-  </div> -->
 </div>
