+++ src/app/calculator/furnaces/furnaces.module.ts
@@ -10,7 +10,6 @@ import { EnergyEquivalencyModule } from './energy-equivalency/energy-equivalency
 import { EnergyUseModule } from './energy-use/energy-use.module';
 import { O2EnrichmentModule } from './o2-enrichment/o2-enrichment.module';
 import { EfficiencyImprovementModule } from './efficiency-improvement/efficiency-improvement.module';
-import { PreAssessmentModule } from './pre-assessment/pre-assessment.module';
 
 @NgModule({
   imports: [
@@ -20,8 +19,7 @@ import { PreAssessmentModule } from './pre-assessment/pre-assessment.module';
     EnergyUseModule,
     O2EnrichmentModule,
     EnergyEquivalencyModule,
-    EfficiencyImprovementModule,
-    PreAssessmentModule
+    EfficiencyImprovementModule
   ],
   declarations: [
     FurnacesComponent
