+++ src/app/calculator/utilities/pre-assessment/pre-assessment-graph/pre-assessment-graph.component.spec.ts
