+++ src/app/calculator/utilities/utilities.component.html
@@ -5,6 +5,20 @@
         <h2 class="mb-1">General Calculators</h2>
         <div class="row calc-item-row">
           <div class="col-1">
+            <img src="assets/images/assessment.png" class="ico-size3">
+          </div>
+          <div class="col-11">
+            <a (click)="showTool('pre-assessment')">Pre-assessment / Screening Calculator</a>
+            <p>
+              This calculator allows the user to enter either energy use information gathered from meters or design specifications (for
+              fuel, electrotechnology or steam furnaces) to calculate energy use for multiple furnaces. This allows users
+              to compare the relative energy use of furnaces within a facility to prioritize energy evaluations and efficiency
+              improvements.
+            </p>
+          </div>
+        </div>
+        <div class="row calc-item-row">
+          <div class="col-1">
             <img src="assets/images/unit-converter.png" class="ico-size3">
           </div>
           <div class="col-11">
@@ -57,4 +71,8 @@
 
 <div class="stand-alone-calc tab-content" *ngIf="selectedTool == 'cash-flow'">
   <app-cash-flow></app-cash-flow>
+</div>
+
+<div class="stand-alone-calc tab-content" *ngIf="selectedTool == 'pre-assessment'">
+  <app-pre-assessment></app-pre-assessment>
 </div>
