+++ src/app/core/core.module.ts
@@ -35,7 +35,7 @@ import { ReportRollupModule } from '../report-rollup/report-rollup.module';
 import { VersionModalComponent } from '../version-modal/version-modal.component';
 import { UpdateModalComponent } from '../update-modal/update-modal.component';
 import { AcknowledgmentsPageComponent } from "../acknowledgments-page/acknowledgments-page.component";
-import { PreAssessmentModule } from '../calculator/furnaces/pre-assessment/pre-assessment.module';
+import { PreAssessmentModule } from '../calculator/utilities/pre-assessment/pre-assessment.module';
 import { MeasurComponent } from '../landing-screen/measur/measur.component';
 import { OpeningTutorialComponent } from '../tutorials/opening-tutorial/opening-tutorial.component';
 import { WindowRefService } from '../indexedDb/window-ref.service';
