+++ src/app/phast/phast.module.ts
@@ -26,7 +26,7 @@ import { PhastResultsService } from './phast-results.service';
 import { EnergyCostsComponent } from './energy-costs/energy-costs.component';
 import { ConvertPhastService } from './convert-phast.service';
 import { PhastCalculatorTabsComponent } from './phast-calculator-tabs/phast-calculator-tabs.component';
-import { PreAssessmentModule } from '../calculator/furnaces/pre-assessment/pre-assessment.module';
+import { PreAssessmentModule } from '../calculator/utilities/pre-assessment/pre-assessment.module';
 import { O2EnrichmentModule } from '../calculator/furnaces/o2-enrichment/o2-enrichment.module';
 import { EfficiencyImprovementModule } from '../calculator/furnaces/efficiency-improvement/efficiency-improvement.module';
 import { EnergyEquivalencyModule } from '../calculator/furnaces/energy-equivalency/energy-equivalency.module';
