+++ src/app/report-rollup/phast-rollup/phast-rollup-pre-assessment-table/phast-rollup-pre-assessment-table.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit, Input } from '@angular/core';
 import { graphColors } from '../../../phast/phast-report/report-graphs/graphColors';
 import { Settings } from '../../../shared/models/settings';
 import { Calculator } from '../../../shared/models/calculators';
-import { PreAssessmentService } from '../../../calculator/furnaces/pre-assessment/pre-assessment.service';
-import { PreAssessment } from '../../../calculator/furnaces/pre-assessment/pre-assessment';
+import { PreAssessmentService } from '../../../calculator/utilities/pre-assessment/pre-assessment.service';
+import { PreAssessment } from '../../../calculator/utilities/pre-assessment/pre-assessment';
 import { MeteredEnergy } from '../../../shared/models/phast/meteredEnergy';
 import { IndexedDbService } from '../../../indexedDb/indexed-db.service';
 import { DirectoryDbService } from '../../../indexedDb/directory-db.service';
