+++ src/app/report-rollup/phast-rollup/phast-rollup-print/phast-rollup-print.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Settings } from '../../../shared/models/settings';
 import { Calculator } from '../../../shared/models/calculators';
-import { PreAssessment } from '../../../calculator/furnaces/pre-assessment/pre-assessment';
 
 @Component({
   selector: 'app-phast-rollup-print',
@@ -14,9 +13,6 @@ export class PhastRollupPrintComponent implements OnInit {
 
   @Input()
   calculators: Array<Calculator>;
-
-  preAssessments: Array<PreAssessment>;
-  showPreAssessment: boolean = true;
   constructor() { }
 
   ngOnInit() {
