+++ src/app/report-rollup/phast-rollup/phast-rollup.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { PhastResultsData } from '../report-rollup.service';
 import { Settings } from '../../shared/models/settings';
 import { Calculator } from '../../shared/models/calculators';
-import { PreAssessment } from '../../calculator/furnaces/pre-assessment/pre-assessment';
+import { PreAssessment } from '../../calculator/utilities/pre-assessment/pre-assessment';
 @Component({
   selector: 'app-phast-rollup',
   templateUrl: './phast-rollup.component.html',
